ディレクトリ構成の変更：ルートディレクトリに投稿一覧ページを設定
